unify effective state reads through vtable_read
